requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
mysqlclient>=2.2.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.5
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
python-dotenv>=1.0.0
django-db-connection-pool>=1.2.4
sqlparse>=0.4.4
